+++ rentACar/src/main/java/com/turkcell/rentACar/business/requests/Invoice/CreateInvoiceRequest.java
@@ -27,6 +27,6 @@ public class CreateInvoiceRequest {
 
 	private Double totalPrice;
 
-	private int totalRentDay;
+	private Long totalRentDay;
 
 }
